# Use lazy reduction divide_and_round_q_last_ntt_inplace.