* Work for pi < 2^60 moduli.
* Obtain 26% -- 37% performance gain in CKKS.rescale.
* Obtain 10%~ performance gain in CKKS.keyswitch

Comparing CKKS.rescale 

|degree|#moduli|before(us)|after(us)|speedup|
|---|----|---|----|---|
|32768|15|25099|17775|1.41x|
|16384|8|6085|4436|1.37x|
|8192|4|1456|1069|1.36x|
|4096|2|299|237|1.26x|