@WeiDaiWD 2020-05-28
Thanks for you contribution. This is truly awesome. One quick question, what compiler was used for this benchmark?