@fionser 2020-05-29
@WeiDaiWD 
MBP 2018, 2.3 GHz Intel Core i5