@fionser 2020-05-29
@WeiDaiWD It seems on gcc version 7.5.0 (Ubuntu 7.5.0-3ubuntu1~18.04), an 
`internal compiler error: in tsubst_copy, at cp/pt.c:14588` pops out.